+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 Changelog
 =========
-## Unreleased
+## 1.0.0-rc.2 - 2018-03-27
 ### Added
 - `TransformFilter::$matchCallback` callable to handle custom logic in determining of a transformer should be applied.
+- `TransformFilter::$transformEmpty` to easily set whether empty response data should be sent to the transformer.  
+ Defaults to `false`.
 
 ## 1.0.0-rc.1 - 2018-03-27
 ### Changed